#26744 "Refactor PullHandler implementations to represent clearly defined strategies
This commit refactors PullHandler and its subclasses, namely ContentTypePullHandler, LanguagePullHandler, and SitePullHandler. The original structure was redefined, shifting from an interface to an abstract class.
The transformative intent is to reshape PullHandler into a true strategy pattern implementation. Subsequently, CustomPullHandler was deleted and its responsibilities were assumed by the abstract PullHandler class. Following this, all implementations of PullHandler have been modified to specifically extend this class, ensuring more coherent structure and clarity relating to their individual functionalities."
